pyproject.toml
[project]
name = "hindusthan"
version = "0.1.0"
description = ""
authors = [
    {name = "robiulsunnyemon",email = "robiulsunyemon@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "fastapi (>=0.122.0,<0.123.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "beanie (>=2.0.1,<3.0.0)",
    "motor (>=3.7.1,<4.0.0)",
    "passlib[argon2] (>=1.7.4,<2.0.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "requests (>=2.32.5,<3.0.0)",
    "google (>=3.0.0,<4.0.0)",
    "google-auth (>=2.43.0,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "hindusthan", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
